Bypass osin.Client when loading AuthZ data (#1)
The problem was the osin.Client is an interface with 4 methods. But
the client-key in the mongo data is actually a struct which has the
fields of osin.DefaultClient.

So when golang tries to coerce the mongodb data from BSON format into
an osin.AuthorizeData struct, it sees the "client" key and tries to
coerce that into an osin.Client interface when instead it should be
coercing that into an osin.DefaultClient struct.

The Client: new(osin.DefaultClient) declaration on line 80 basically
means "coerce information in the 'client' key into an 
osin.DefaultClient struct".